roachpb: unskip TestReplicaUnavailableError
This test needs to live in string_test.go for reasons that have to do
with how we inject `keys.PrettyPrint` into `roachpb`. In the bazel
build, this is only done for `strings_test.go` (for reasons I don't
entirely understand, but I assume it is hard to override gazelle
only partially, so we picked a file that would be dep'ed manually
to contain the problem to this one file).

Repurposing #75108 to track making this workaround more obvious. It took
me a few hours to figure this all out, and it'll look similarly
confusing to the next person.

Release note: None
